hermes-daemon/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.2
discord.py==2.4.0
python-dotenv==1.0.1
